add kapasitas per hari

+++ database/migrations/2025_09_08_145526_create_production_t_data1.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('KTEXT')->nullable();
             $table->string('ARBID')->nullable();
             $table->string('KAPID')->nullable();
-            $table->integer('KAPAZ')->nullable();
+            $table->float('KAPAZ')->nullable();
             $table->string('VERID')->nullable();
             $table->string('KDAUF')->nullable();
             $table->string('KDPOS')->nullable();
